VM Provisioning (#3)
This branch adds support for provisioning Grafana to a VM in GCP along with a scheduled job that cleans up instances, firewall rules, and static ips created by Bench.

It also reworks the internals of how the project is organized and splits build, provision, test into separate packages which are then configured in the magefile.

This pushes bench very close to a 1.0 release. Previous versions of Bench are now deprecated and should no longer be used.