Remove unused bench code (#209)
This PR removes legacy code for the original Mage-based architecture and moves dependencies used by the new architecture into the pkg/ directory. At some point in the future we may re-introduce build and provision commands, however, those are currently unused. We may reintroduce this functionality in the future based on a new architecture if required.

---------

Signed-off-by: Pablo Chacin <pablochacin@gmail.com>