add chrome deps back to dockerfile (#603)
This PR adds chrome back to docker image.

Chrome requires system deps. When running playwright install it will give us a notice that we need to install those deps which require sudo access. When installing the deps directly in the image, playwright tests hang, so for now we are just adding chrome back to the image.